refactor: move note to mod.rs, as in original API
